+++ frontend/pasientdata2021/src/components/pages/MapPage.js
@@ -9,10 +9,7 @@ import { useHistory } from 'react-router';
 
 import { useState } from 'react';
 import ArrowButton from '../buttons/ArrowButton';
-import { Fab } from '@material-ui/core'
-import AddIcon from '@material-ui/icons/Add'
-
-import { makeStyles } from '@material-ui/core/styles';
+import FloatingButton from '../buttons/FloatingButton';
 
 const Wrapper = styled.div`
     display: flex;
@@ -28,20 +25,6 @@ const TopContainer = styled.div`
 const MapContainer = styled(MapComponent)`
     height: 100%;
 `
-const FloatButtonContainer = styled.div`
-    position: fixed;
-    top: inherit;
-    margin-top: 20px;
-    margin-left: 82%;
-    color: #7BEFB2;
-`
-
-const useStyles = makeStyles((theme) => ({
-    color: {
-        color: "#7BEFB2",
-        backgroundColor:"#7BEFB2",
-    },
-}));
 
 const SwipeableBottomStyle = {
     width: "inherit"
@@ -60,8 +43,6 @@ const SwipeableBottomContainer = styled.div`
 `
 
 function MapPage() {
-
-    const classes = useStyles();
     const history = useHistory();
     const [open, setOpen] = useState(false)
 
@@ -69,11 +50,7 @@ function MapPage() {
         <Wrapper className="MapPageWrapper">
         <TopContainer className="TopContainer"></TopContainer>
         <MapContainer className="map-container" />
-        <FloatButtonContainer>
-        <Fab className={classes.color} aria-label="add" size="large">
-            <AddIcon style ={{ fontSize: 40, color: "black" }} onClick={() =>history.push("/createtrip")}/>
-        </Fab>
-        </FloatButtonContainer>
+        <FloatingButton className="FloatingButton" onClick={()=>history.push('/createtrip')} />
         <SwipeableBottom style={SwipeableBottomStyle} open={open} setOpen = {setOpen}>
               <SwipeableBottomTop onClick={()=>setOpen(!open)}>
                 <ArrowButton direction="up"></ArrowButton>
